Tests: Move host fixtures to conftest.py
