+++ monkey/tests/unit_tests/infection_monkey/exploit/conftest.py
@@ -19,3 +19,36 @@ def patch_win32api_get_user_name(local_user):
     win32api.NameSamCompatible = None
 
     sys.modules["win32api"] = win32api
+
+
+def _create_host(http_enabled, https_enabled):
+    host = MagicMock()
+    host.services = {}
+
+    if http_enabled:
+        host.services["tcp-5985"] = {}
+
+    if https_enabled:
+        host.services["tcp-5986"] = {}
+
+    return host
+
+
+@pytest.fixture
+def https_only_host():
+    return _create_host(False, True)
+
+
+@pytest.fixture
+def http_only_host():
+    return _create_host(True, False)
+
+
+@pytest.fixture
+def http_and_https_both_enabled_host():
+    return _create_host(True, True)
+
+
+@pytest.fixture
+def powershell_disabled_host():
+    return _create_host(False, False)
