+++ monkey/tests/unit_tests/infection_monkey/exploit/powershell_utils/test_auth_options.py
@@ -1,7 +1,3 @@
-from unittest.mock import MagicMock
-
-import pytest
-
 # from infection_monkey.exploit.powershell_utils.auth_options import AuthOptions
 from infection_monkey.exploit.powershell_utils.auth_options import (
     AUTH_BASIC,
@@ -18,39 +14,6 @@ CREDENTIALS_EMPTY_PASSWORD = Credentials("user2", "", SecretType.PASSWORD)
 CREDENTIALS_NONE_PASSWORD = Credentials("user3", None, SecretType.CACHED)
 CREDENTIALS_LM_HASH = Credentials("user4", "LM_HASH:NONE", SecretType.LM_HASH)
 CREDENTIALS_NT_HASH = Credentials("user5", "NONE:NT_HASH", SecretType.NT_HASH)
-
-
-def _create_host(http_enabled, https_enabled):
-    host = MagicMock()
-    host.services = {}
-
-    if http_enabled:
-        host.services["tcp-5985"] = {}
-
-    if https_enabled:
-        host.services["tcp-5986"] = {}
-
-    return host
-
-
-@pytest.fixture
-def https_only_host():
-    return _create_host(False, True)
-
-
-@pytest.fixture
-def http_only_host():
-    return _create_host(True, False)
-
-
-@pytest.fixture
-def http_and_https_both_enabled_host():
-    return _create_host(True, True)
-
-
-@pytest.fixture
-def powershell_disabled_host():
-    return _create_host(False, False)
 
 
 def test_get_auth_options__ssl_false_with_no_open_ports(powershell_disabled_host):
